Add Stack Navigator to App and Auth
